refactor: 네이밍 변경 및 패키지 구조 변경
