+++ src/main/java/gift/service/dto/KakaoInfoDto.java
@@ -1,4 +1,4 @@
-package gift.util.dto;
+package gift.service.dto;
 
 public record KakaoInfoDto(
         String id,
